poa_backend v0.1.0 POABackend.Auth

This module defines the API for the Authorisation

Functions

activate_user(user)

activate_user(POABackend.Auth.Models.User.t()) ::
  {:ok, POABackend.Auth.Models.User.t()} | {:error, Ecto.Changeset.t()}

This function Activates a user, storing active: true in the database for the given user

authenticate_admin(admin_name, password)

authenticate_admin(String.t(), String.t()) ::
  {:ok, :valid} | {:error, :notvalid}

Authenticates an Admin

authenticate_user(user_name, password)

authenticate_user(String.t(), String.t()) ::
  {:ok, POABackend.Auth.Models.User.t()} | {:error, :notvalid}

This function authenticates a user/password pair

create_banned_token(jwt_token)

create_banned_token(String.t()) ::
  {:ok, POABackend.Auth.Models.Token.t()} | {:error, any()}

Creates a token entry in the banned tokens table. It receives a jwt token in String format. This function will extract the expiration time from the claims and store them in the Database.

create_banned_token(token, expires)

create_banned_token(String.t(), integer()) ::
  {:ok, POABackend.Auth.Models.Token.t()} | {:error, :already_exists}

Creates a token entry in the banned tokens table. It receives the token in String format and the expiration time as Integer.

create_user(user_name, password, active \\ true)

create_user(String.t(), String.t(), Boolean.t()) ::
  {:ok, POABackend.Auth.Models.User.t()}
  | {:error, :already_exists}
  | {:error, Ecto.Changeset.t()}

Registers a user in the system.

deactivate_user(user)

deactivate_user(POABackend.Auth.Models.User.t()) ::
  {:ok, POABackend.Auth.Models.User.t()} | {:error, Ecto.Changeset.t()}

This function Deactivates a user, storing active: false in the database for the given user

generate_password()

generate_password() :: String.t()

Generates a password randomply

generate_user_name()

generate_user_name() :: String.t()

Generates a valid user name randomply

generate_user_name(user_name)

generate_user_name(String.t()) :: String.t()

This function is exported for testing purposes

get_user(user)

get_user(String.t()) :: POABackend.Auth.Models.User.t() | nil

Get a user from the database based in the user name

purge_banned_tokens()

purge_banned_tokens() :: :ok

This function deletes the banned tokens which already expired

remove_user(user)

remove_user(POABackend.Auth.Models.User.t()) :: :ok

Deletes a user from the database based in the given user

token_banned?(token)

token_banned?(String.t()) :: Boolean.t()

Checks if a token is banned or not

user_active?(user)

user_active?(POABackend.Auth.Models.User.t()) :: Boolean.t()

Checks if a user is active

valid_token?(jwt_token)

valid_token?(String.t()) :: Boolean.t() | {:error, :token_expired}

Validates if a JWT token is valid.

valid_user_name?(user_name)

valid_user_name?(String.t()) :: Boolean.t()

Validates if a given user name is valid or not. It is valid if doesn’t exist a user with that name in the database already