poa_backend v0.1.0 POABackend.Receivers.Eth.Stats

This is a Receiver Plugin which is in charge of storing the Ethereum Statistics received from the Agents in a Postgres database. If we want to use it we have to declare it in the Config file, inside the :receivers section, ie:

{:store_eth_stats, POABackend.Receivers.Eth.Stats, []}

This Plugin uses Postgres as a backend, specifically the eth_stats table. Make sure that table exists before using this Plugin.

We also need to subscribe this plugin to :ethereum_metrics metrics in the config file. For that we have to add this line to the list in the :subscriptions section:

{:store_eth_stats, [:ethereum_metrics]}

Functions

handle_inactive(agent_id, state)

This function is called when a Custom Handler detects a client is inactive.

The Custom Handler must to call explicity to POABackend.CustomHandler.publish_inactive/1 and it will publish the inactive message to all the metrics in the system (defined in the config file).

Callback implementation for POABackend.Receiver.handle_inactive/2.

handle_message(message, state)

In this callback is called when the Receiver process receives an erlang message.

It must return {:ok, state}.

Callback implementation for POABackend.Receiver.handle_message/2.

init_receiver(args)

A callback executed when the Receiver Plugin starts. The argument is retrieved from the configuration file when the Receiver is defined It must return {:ok, state}, that state will be keept as in GenServer and can be retrieved in the metrics_received/3 function.

Callback implementation for POABackend.Receiver.init_receiver/1.

metrics_received(arg1, from, state)

This callback will be called every time a message to the subscribed metric type arrives. It must return the tuple {:ok, state}

Callback implementation for POABackend.Receiver.metrics_received/3.

terminate(state)

This callback is called just before the Process goes down. This is a good place for closing connections.

Callback implementation for POABackend.Receiver.terminate/1.